Decouple some tools from the disk-based index.
You can now pass a TraceParams structure to TarmacUtility, which will
be passed in turn to the indexer. Flags in TraceParams cause the
indexer to leave out some of the data in the index. In particular, you
can suppress indexing of memory (leaving only registers), and you can
suppress generation of the layered range tree of call depths. The idea
is that clients can set these flags if they don't need those features,
and avoid spending time on constructing them and disk space on storing
them.

I've set record_memory = false in five existing tools, which work
entirely from the call stack and register contents. No tool yet sets
record_calls = false, but it was an easy flag to add.

An immediate effect of that is that those five tools now no longer use
a disk-based index file at all: their memory needs are now much
smaller, so they don't really _need_ a disk index, and unconditionally
getting rid of it means that we don't have to introduce a system of
flags in the index file indicating what parts of it are missing (and
then have tools regenerate it in a more complete form if needed).

Perhaps later that flag system might be necessary after all, in which
case tarmac-callinfo and friends could go back to using the index (or
at least using an existing one if it's present). But this was the
simple thing for now.