AVLDisk: support a reference-counting mode.
The aim of this is to enable the Tarmac indexer to operate in a
streaming mode, in which a client processes each node of the
sequential order tree as it's generated, and never looks back at
history. Therefore, there's no need to store every memtree permanently
and gradually increase the space usage: instead, we can just keep
track of the _current_ state of memory.

However, it's still potentially useful for a client to be able to see
the states of memory before and after each event. So this commit
enhances the AVL tree system so that it can permit node-sharing
between trees using a more dynamic system than the fixed 'high-water
mark' rule: each node has a reference count indicating how many other
nodes (or distinct tree roots) currently refer to it. This allows the
data structure to detect when a node is shared between two trees,
because it (or one of its ancestors) has a reference count > 1. In
that situation you clone the node and modify the new copy, just as in
the existing HWM system - but a side effect is that reference counts
in the existing tree decrease, so the previously immutable nodes end
up mutable once nothing else is using them any more.

No functional change: the new mode is not yet used by anything.

However, I've put the refcount field into the data structure
unconditionally (it was too hard to do it conditionally!), so this is
an index-file format change. Bumped the version number accordingly.