Break up parse_tarmac_file() into subroutines.
The idea is to lay groundwork for greater flexibility, in particular
having the trace parsing and indexing run incrementally so that a
client can process the results in line with the parse.

It might also enable a mode in which the trace is read once without
tracking memory contents to generate the call stack analysis, and then
a second time _with_ memory contents tracked (in the new
non-persistent AVL tree mode) so that the call information is
available as well as the current memory state.

No functional change.