Add CI with Ubuntu 22.04.
And refactor so that gcc / clang testing are done in the same workflow.
Ideally, they should be able to share the same environment, that is the
machine once the additional dependencies have been installed, but I do
not know how to do that yet.

Building is now done with 2 jobs in parallel, as this is the number of
cores available to us.

And update the checkout action to v3 as v2 is now tagged as deprecated.