tarmac-browser: make the coloured display optional.
The new --colour and --no-colour options toggle whether tarmac-browser
displays its terminal UI in colour.

There's a new set of curses attribute settings for when the UI is not
using colour. In that setting, window status lines are set to reverse
video, so that they give the same visual separation between the
windows that the colour display achieved by changing the background
colour.

Other than that, the non-colour base attributes are the same as the
colour ones, for the moment. But they're stored separately in the
ATTRLIST list macro, so it's easy to change them later if anyone has a
stronger opinion than me about which things should be bold.

In the absence of either command-line option, the default is still to
use colour, but you can set NO_COLOR in the environment to override
that, as recommended by https://no-color.org/ .

We also now check if the terminfo system reports that colour support
is available at all, and switch to the new monochrome attribute set if
it's not, so that users in that situation will now automatically get
the improved reverse-video status bars.