Fix compile warning about getenv on Windows.
Commit e9b24783c2be2a4 introduced a check in the curses-based Tarmac
browser for the environment variable NO_COLOR, by calling getenv() at
the start of main(). Unfortunately this provokes a deprecation warning
in MSVC, which doesn't like getenv() on unspecified security grounds,
and suggests the use of _dupenv_s instead.

In this case I'm sure that getenv() is harmless. As far as I can tell
from the MSVC library documentation, the complaint about getenv has to
do with race conditions if another thread is _modifying_ the
environment, because it returns a reference in place to the string
value of the variable. So calling getenv in a program that never
starts a second thread at all (and even if it did, hasn't done it
_yet_ at that point in main()), _and_ which never modifies its own
environment, is safe regardless. But there's no way to annotate just
one call to getenv() to suppress the warning.

Fixed by writing a platform-specific function to get an environment
variable. Overkill, perhaps, but it makes the warning go away.