[VCD] Optionally use the instructions timestamps from the tarmac trace.
This allows to an easier matching with RTL-level simulation traces. This
is however not enabled by default as not all tarmax traces are from
RTL-level simulations, plus the VCD trace that we generate is synthetic,
i.e. we try to present to our user the data in a convenient way without
a complete view of the information, e.g. in the case of LDM/STM we (at
best) guess the sequence of transfers on the data bus.