Fix format-string bug in config file parse errors.
When tarmac-gui-browser parses its user configuration file, two of the
possible parse errors are reported by constructing a std::string and
then calling reporter->warnx(message.c_str()). But warnx() is a
printf-style function (following the one in <err.h>), so it's unsafe
to pass an arbitrary string as its _first_ parameter. If you wrote an
invalid config directive with a %s in it, you could cause the browser
to crash.

Fixed by the usual safe idiom of warnx("%s", message.c_str()).

(I don't think this is a _security_ issue, because the configuration
file isn't under hostile control. But it's obviously a bug.)