Introduce new ParseParams structure.
This structure replaces the previous 'bigend' flag that was passed to
TarmacLineParser at construction time. No functional change: currently
the only thing it contains is that same 'bigend' flag.

Rationale: I'm going to add a new parameter shortly, which
disambiguates parsing a trace that didn't provide enough information.
This is also a parameter that affects how to parse the trace, so it
seems sensible to put it alongside 'bigend' as members of a common
structure type.

This refactoring also fixes the slight laziness in which HighlightedLine
couldn't be bothered to receive the big-endian flag because it didn't
matter for highlighting. The next flag _will_ be needed for highlighting,
so we have to make sure to pass the parameters structure to
HighlightedLine's constructor.