Parser: handle traces that are implicitly Thumb only.
This fixes the other half of issue #13. The trace snippet in that
issue omits several fields from the usual format of the IT instruction
record line, of which the important one to us is the instruction set
state. But the trace comes from a Cortex-M0, which only _has_ one
instruction set state, namely Thumb.

I don't think it's sensible to assume that _any_ trace file that omits
the state field must be Thumb. (For all I know, there might be a
similar format emitted from some AArch64-only system.) So instead I
think we just have to require the user to disambiguate on the command
line. That's a shame, but it can't be helped. And users already have
to do that for big-endian traces, so it's not even 100% unprecedented.

I've added an --implicit-thumb flag to all the tools that read trace
files, and had it set the new 'iset' field of ParseParams. The parser
will reject IT lines in this truncated format by default, on the
grounds that it doesn't know what instruction set state to put in
them; but if you tell it there is only one, then it will accept that
type of line.

The same flag is also supported by parsertest.txt, and there's a
separate test of the parser containing sample trace lines affected by
the flag.

Like the big-endian flag, the new 'implicitly Thumb' flag wants to be
recorded in the index file header. So this is an index-file format
update, and I've bumped the magic number.