Parser: handle ES records in which insn fetch failed.
From a couple of sources recently I've received samples of trace in
which an ES-style instruction record has the instruction bit pattern
replaced with "--------" and no disassembly. Apparently this can arise
if the instruction is being fetched from ECC memory, and the ECC
system detects an inconsistency.

In the trace snippets I saw, this wasn't a fatal condition - no
instruction fetch abort was taken. It's error _correcting_ memory, not
just error detecting: given time, it can get itself back into a
consistent state by doing the algebra. But the failed first attempt,
and the pause while the memory sorts itself out, show up in the trace,
and need to be handled by our trace parser.

The obvious thing would be to reclassify an ES line of that type as a
TextOnlyEvent, since it doesn't actually indicate any change in the
CPU's state: the next instruction fetch will come from the same
address again, and this time, succeed. However, there's a subtle
problem with that: because the _timestamp_ of the line has advanced
from the previous instruction, it still begins a new node in the trace
index's sequential order tree, and if that node only contains
TextOnlyEvents then it won't have a valid PC at all.

So, instead, I've changed the boolean 'executed' field of
InstructionEvent into a three-state enum, with the new IE_FETCHFAIL
state indicating a line of this type. That way, we can still record
all the rest of the metadata that goes with an _attempted_ instruction
execution, like the PC and the iset state.

IE_FETCHFAIL is currently treated identically to IE_CCFAIL by all the
tools. I keep them separate partly so that IE_FETCHFAIL can signal not
to even _look_ at the 'instruction' field containing the encoding, and
also, because they have different effects on the following PC value:
in IE_CCFAIL you still expect the PC to advance to the next
instruction, but IE_FETCHFAIL is weirder.