Rename TraceParams to IndexerParams.
I'm about to introduce another parameters structure, so I want the two
to have names that clearly communicate what they do. Both are about
trace files, so calling either one TraceParams is silly; the important
thing is that this structure describes how we build an _index_ from
the trace (in particular, what things we leave out of that index).

No functional change.