Add a --load-offset option for debugging PIC images.
This has been a FIXME in the code for ages: if an ELF image contains
position-independent code, then you expect that sometimes it will be
loaded into memory at a location that doesn't match its segment and
symbol addresses. In that situation it's useful for TTU to be able to
map the addresses in the trace to symbols in the ELF file taking
account of the different load address. So this option lets you tell it
the offset between the two addresses.

The implementation of this is all simple. The hardest part was
documenting clearly which way round the offset works, i.e. that it's
the offset _from_ an address in the image _to_ the address where the
corresponding thing was actually loaded.