Permit consecutive LD and ST records to omit their keyword.
Fixes #15, whose reporter sent in a snippet of trace from a recent
Neoverse core (but didn't say which one).

Recognising when an LD or ST line has been continued in this way is
hard to do robustly, because the first token on a continuation line is
an address in hex, and if that happens to contain no digits in {a-f}
then the two could be confused.

So instead I've opted for an 'offside rule', making whitespace
significant. If the first token on the line after an LD or ST is
aligned with the address token from the previous line, then it's taken
to be a continuation line.