i18n prep: cmake checks for the libintl library.
This is the start of changes to make TTU able to accept translations
of all its messages, via the GNU gettext tools and libintl library,
and their system of marking translatable strings inside source files
with _("...").

The libintl code has to be optional, because Windows builds can't use
it as far as I know. So I provide a wrapper around <libintl.h> that
conditionally includes it, and defines _() appropriately.

No functional change.