i18n prep: introduce a gettext_setup() function.
This is suitable for calling at the start of each main program, and
performs GNU gettext's recommended sequence of setup function calls.

The string-valued macros PACKAGE and LOCALEDIR must be passed in from
the cmake system. The package name can be automatically retrieved from
the variable where CMake's 'project' command saved it. The locale
directory defaults to the appropriate subdirectory of
CMAKE_INSTALL_PREFIX, but I've also made it reconfigurable as a CMake
cache variable, just in case.

The new function is called at the start of main() in all the
user-facing tools, but not in the internal test programs such as
parsertest, exprtest, avltest etc.