Windows: fix trailing \0 in get_environment_variable.
_dupenv_s apparently returns the length of the environment variable's
value _including_ the trailing NUL, so when we made a std::string out
of it using the returned length field, we got the NUL included as part
of the semantic string contents.

I can't actually see that behaviour documented in the manual page for
_dupenv_s, so I've been a bit cautious about undoing it: we actually
check the string really does end in a \0 before trimming it off.