Converter from gettext to Windows message resources.
The gettext system is good for translatable programs on Unix, but TTU
also runs on Windows, which doesn't have a sensible place to keep
directories full of translation tables.

Windows's own system for translation is to include a binary resource
in the program, containing the messages for all languages indexed by a
numeric id per message, and then use the 'FormatMessage' API function
to look up a message in that resource section in the currently
configured language.

That system is more intrusive to your source code: you lose the
ability to have the original text of a message right next to the code
that relates to it, and instead, everything has to turn into an enum
of message codes with even the original language's text being kept in
a file somewhere else. I think gettext's approach of using the
original string as the lookup key is more ergonomic.

So I've stuck with the gettext idiom in the source code, and developed
a mechanism for generating Windows message resources _from_ the binary
files output by the gettext message compiler. Those files have a nice
easy format (apart from an undocumented and optional hash table
component which we don't need anyway), so it's no trouble to read one
and construct the binary format of a Windows message resource.

Of course, if the source code uses the input text as its lookup key,
but FormatMessage wants an integer index, then we also need a lookup
table in the other direction, for translating a source string to an
index. This script generates a matching one of those too.

In this commit, I just commit the script itself. Nothing uses it yet.

Note on testing: unfortunately, Windows's system of language IDs
doesn't permit developers to randomly make up extra ones for test
purposes. So I can't assign permanent LCIDs to the 'wide' and 'rot13'
test translations. Instead, I have to test them by bodging the
conversion script to pretend they're a real language that has an LCID.