Add cmake setup for building translations.
po/CMakeLists.txt refers to each subdirectory containing a
translation; the subdirectory in turn has a call to a cmake gettext
function that does the right thing, mentioning the name of the
language in particular.

This also arranges that the cmake install target will install the
translation into the share/locale subdirectory of the install prefix.
That's also the path baked into the binaries (via the #define of
LOCALEDIR in cmake.h.in, which will be set based on the install prefix
when that's turned into the real cmake.h in the build directory), so
once you've run 'make install' (or 'ninja install' or whatever), the
tools will be able to find message locale files built by this step.