Remove spurious "tarmac-browser:" in an error report.
When index.cpp sees a parse error on the last line of the trace file,
it reports a special message, which starts with it announcing the
tool's identity as "tarmac-browser".

But index.cpp is shared between all the tools - it's been a long time
since tarmac-browser was the only binary built by this code base.
Also, the current error reporting system writes that message _after_
the filename and line number of the problem, which would be a strange
place to write the program name anyway.

If the program name is to be printed at all in the current reporting
system, better that CommandLineReporter should uniformly prefix it to
everything it prints. But I think in this case it's not even
necessary: the trace file name is enough context.