Windows: support an LCID_OVERRIDE env variable for testing.
For testing a translation, it's awkward to _really_ configure a
Windows machine into a different language. Much easier to tell the
program under test to just set its own locale differently.

So now, if you've made (say) a French translation and want to test it
on a non-French-configured Windows machine, you can set
"LCID_OVERRIDE=0x40C" in the environment, and tools in this code base
will pick it up and reset their thread locale.

(It would be nice if there was a _standard_ variable for this, like
LC_FOO or LANG or LANGUAGE on Linux, but I don't know of one.)