i18n prep: introduce a stopgap format function.
As part of making strings translatable, we're going to have to move
away from the trad C++ style of building up messages out of a lot of
separate string literals and interleaved pieces, either via ostream <<
syntax or via string concatenation. Instead, messages are going to
need to have overall format strings with a syntax for interpolation.

The nicest way to do this in C++ would be to use C++20 std::format,
but it's not yet supported in any compiler available on Ubuntu 22.04.

Another approach would be to use printf format strings, and have a
shim function with a printf-like API that returns a std::string. But
that would have all the usual disadvantages of C printf in C++:
avoidable type-unsafety, and having to suffix all your std::string
format parameters with .c_str().

So instead I've written a tiny formatter that supports a subset of the
std::format format-string syntax. I plan to use this until we can
start using real std::format in its place, and then we should be able
to switch over without having to change any existing format
strings (which are also keys in the translation file, so it's a good
idea to avoid changing them unnecessarily).