FormatMessage-based lookup for translated messages.
This fills in a Windows branch of the #if in intl.hh, and some
corresponding code in the Windows platform module. Now there's a
function that the Windows tools can call, with the same API as
gettext(3), but not actually using the real gettext library. Instead,
it's implemented in this code base: it uses the reverse translation
table generated by mo2msgres.py to turn the input string into an
index, and then FormatMessage to look up that index for the current
language.