Try to account for wide characters in terminals.
When argparse.cpp displays help text that's supposed to line up nicely
in two word-wrapped columns in a fixed-pitch terminal, it needs to
take account of the fact that some characters (particularly East
Asian, but not only) take up two character cells, and some (combining
characters and zero-width administrative things like joiners) take up
none.

The library functions wcwidth() and wcswidth() know how to figure this
out, so we test for their availability at cmake time. If we can't get
them, there's not much we can do about it (short of importing the
relevant data from the Unicode standard and reimplementing them).

This should also mean that Unicode combining characters are correctly
identified as occupying _zero_ character cells, so that text using
those will also wrap to a sensible width.