Add a Python script for running xgettext.
This script knows how to run the 'xgettext' utility to create or
update a message templates file for this code base.

You have to remember a couple of fiddly options, and also pass it the
right set of source files to scan. Since we have to exclude the
samples directory, the latter isn't quite as trivial as 'git
ls-files'. So I decided it was easiest to make a script which can be
run without arguments, and knows how to do the right thing.