Fix building with libintl on macOS.
CI pointed this out after I pushed Friday's patch series: I had used
the cmake FindIntl module to test whether libintl was available, but
having found that it was, I'd assumed that it wouldn't require any
actual change to the build commands.

On Linux, where the header file lives in /usr/include and the library
functionality is integrated into libc, this is indeed true. But it's
not true everywhere. So we still have to remember to _use_ all the
other data returned by FindIntl.

This is less painful than it could have been, because I've made a new
'standard_target_configuration' function which adds standard stuff to
every executable target we build, and folded in the existing statement
that links them against libtarmac. So it's not a _new_ function call
per target - it mostly replaces one of the existing ones.