Provide a manual cmake flag to turn off libintl.
This is likely to be useful for people building out-of-tree clients of
libtarmac: partly because the sudden dependency on libintl might be
inconvenient for them, and also because the use of libintl might not
even work if their own application doesn't integrate libintl and
provide message translations.

(It's possible that to support out-of-tree libtarmac clients fully, we
might need to separate the messages in this project into a set for the
library and a set for the applications. But this switch is surely
_also_ useful.)