Make add_this_po a macro, not a function.
The function add_this_po tries to modify the variable 'mo_files' in
its parent scope, intending to append its output file to the list
variable used by po/CMakeLists.txt.

But it's both a cmake function, _and_ running inside a cmake
add_subdirectory() relative to po/CMakeLists.txt. And in cmake, both
of those operations establish a new variable scope. So add_this_po
needs to reach up _two_ levels to affect the variable it really wants,
and set(... PARENT_SCOPE) only reaches up one.

The simplest answer is to turn it from a function into a macro,
because macros don't have an internal variable scope, so now the set()
only has to reach one level up.

As a result of this, the 'mo_files' list variable was ending up empty,
when it should have contained a list of files to be passed to the
Python script that builds Windows message resources.