Fix a tricky edge case of LD/ST continuation.
I received a snippet of trace recently generated by a Cortex-M33 RTL
simulation, which omitted the LD or ST keyword from a two-line load or
store in the same manner that commit 49cf4c95ba12b98 was intended to
support (though all the previous cases I'd seen were in AArch64). With
a 32-bit address, it's that much more likely that the hex address at
the start of the continuation line might happen not to have any
alphabetic hex digits in, i.e. it might be composed entirely of
decimal digits. So then you have to be careful not to mistake it for a
timestamp!