Fix line wrapping in --help.
Some of the longer help messages were being mis-wrapped in a way that
involved starting a new line earlier and earlier, so that you ended up
with the tail of the message having just one word per line.

This was apparently introduced in commit ed5884db20c770a, quite
recently: when it became harder to compute the line length during
wrapping (because it's done with wcswidth instead of just
string::size()), I started keeping running track of it, and took the
shortcut of updating it after each word by setting it to
'candidate_len' which was computed at the start of the loop iteration.
But in the case where we started a new line after candidate_len was
computed, its previous value became invalid. Now we recompute it.