Fix character encoding in Windows consoles.
While testing out the new French translation of the messages, I found
that they were mis-encoded on Windows, in a way that suggested that
text encoded in ISO 8859-1 or Win1252 had been interpreted as if it
were CP850.

On the test Windows machine, that is indeed exactly what happened,
because GetACP() returns 1252; GetOEMCP() returns 850; and apparently
the console is using the latter, which is easy to check by printing a
test sequence of all high-bit-set byte values.

The char version of FormatMessage() has no flag to ask for output in a
particular code page, so I've had to do it by hand, explicitly calling
FormatMessageW() and then converting afterwards. This shouldn't be too
slow, because the results are cached per message index anyway.

The target character set is retrieved using GetConsoleOutputCP()
rather than GetOEMCP(), because consoles' character sets can be
reconfigured. Also, this has required me to add a boolean flag to
gettext_setup() to tell it whether the application is talking to a
console at all – otherwise you'd get CP850 strings being sent to the
GUI by tarmac-gui-browser as well, and the same problem would happen
in reverse.

Of course, the most orthodox Windows response would be to start moving
towards storing strings internally in Unicode. But that has its own
portability issues (Unix is less keen on wchar_t), and also, it's a
big enough piece of work that it wouldn't be a good idea to try to do
it 'in passing' to fix this.