Handle overlong 'cpsr' in Fast Models trace.
The Fast Models style of Tarmac uses a register update of 'cpsr' to
mean more than just the literal CPSR; it also includes a few extra
bits (as if CPSR were 64-bit rather than 32) that architecturally live
in different system registers.

In this commit, we don't _use_ the extra bits, we just tolerate their
existence without throwing a parse error.