Rename the French locale to 'fr'.
I noticed that Arnaud's .po file was in a 'fr_FR' directory but had
'Language: fr' in its header. It turns out that that's because msginit
truncated the locale description on purpose, because it has a list of
two-part locale definitions that it promotes to one-part, and in this
list, fr_FR is the default kind of fr.

This seems to be the convention followed by almost all other Linux
software with a French translation: in a straw poll of my own
/usr/share/locale/*/LC_MESSAGES, the 'fr' directory has almost all the
entries, and 'fr_FR' contains only a couple of files, both of which
are stubs containing no actual text.

So I think if the consensus is that strong, we should probably follow
it too. Accordingly, I've renamed the fr_FR directory to just fr (so
that now it matches the Language: header), and added some text to
index.rst suggesting how to test whether msginit thinks this ought to
be done for any other translations.