mo2msgres.py: add 'fr' to the LCID mapping list.
This arranges that the new French translation is incorporated into the
Windows message resource.

In contrast to the previous commit's message, here we _do_ have to
take a position on which flavour of 'fr' we mean, because as far as I
can see, the FormatMessage lookup system doesn't do any kind of
fallback between them: if we set our translation to be 0x0C for plain
'fr' then it won't be selected on a machine localised to 0x040C (fr_FR).