CMakeLists.txt: fix misspelling of message(FATAL_ERROR).
message(FATAL "...") doesn't function as a fatal error at all. It
just prints "FATAL" prepended to the error string and carries on.
Corrected the spelling so that it makes cmake actually fail.

(I observed this when the check for off_t actually failed on a test
Windows build. But the failure had nothing to do with off_t, and also
seems to be intermittent - some kind of problem at link time inserting
a file manifest, probably due to file sharing issues or similar - so
there isn't also a fix I can make in the actual off_t logic.)