Report an error message for an empty trace file.
In the previous commit message I mentioned that the effect of failing
to notice the trace file didn't exist was to abort at the end of trace
construction. But that abort can also be reached if the trace file
_does_ exist and is empty. Specifically, I think the problem arises if
the main seqtree has no entries in it at all. So now we check for that
condition, and terminate indexing with an error.

This also needed a separate fix in the GUI browser, which failed an
assertion _during_ indexing due to using the file size as the limit of
its progress bar, and an assertion inside wxWidgets checks that limit
is not zero, presumably to avoid that division by 0 that would
otherwise occur. So in that code I've ensured progress_max > 0 always,
which means tarmac-gui-browser now survives long enough to get to the
centralised error at the end of indexing.