tarmac-profile: force at least one space between columns.
If you profile a 64-bit image whose code addresses are at least
0x1000000000, then each function address occupies the whole of the
12-character 'Address' column in the tabular output, or more. So the
number in the 'Count' column is concatenated to the end, making the
display ambiguous and hard to read.

Now we reliably print at least one space between the numbers, even if
they're wide enough to make the alignment go wonky.