Call heuristics: defuse 'recent lr update' after a branch.
This protects against misidentifying an internal branch within a
function as a potential call on the strength of a sequence such as

  callee pops LR
  callee returns to LR
  caller tests return value
  caller conditionally branches

in which the final conditional branch would previously have been a
PendingCall candidate on the grounds that LR had been written <8
instructions previously. But I think that because another branch
(the RET) happened _between_ the LR write and the conditional branch,
that disqualifies the latter from being a call with that LR as its
return value.

The same thing can also happen to a conditional branch just after a
call, because that too is a moment where LR has just been written. But
that version of the problem only triggers if the call instruction and
the callee are very close together, because of the 'within 64 bytes'
clause in the heuristic analysis. However, that can happen when a
function recursively calls itself - which caused 7 of our existing
regression tests to fail, because the call tree analysis of
tests/quicksort.tarmac had changed, because that error was happening
in the recursive calls to quicksort!

(An easy-to-observe reason why the new expected output for those tests
is better: the tarmac-calltree output with symbols listed some calls
to quicksort as having address 0x8038 and some as 0x8054, which is
suspicious, because surely a function should have only one address. In
fact 0x8038 is the real address of quicksort() in that test trace, and
0x8054 is an internal branch label. Now all the calls to quicksort
list the former.)