Replace -DDEBUG_CALL_HEURISTICS with a runtime option.
When the matching up of calls and returns isn't working, it's awkward
to have to recompile the tools with a special #define to fix it. Also,
having the same facility available on demand in tools built the
standard way means I can make a set of regression tests for the call
matcher, by capturing the diagnostics and checking them.